LayoutTests/imported/w3c/web-platform-tests/css/css-nesting/nested-declarations-matching-expected.txt

A1
A2

PASS Trailing declarations apply after any preceding rules
PASS Trailing declarations apply after any preceding rules (no leading)
PASS Trailing declarations apply after any preceding rules (multiple)
PASS Nested declarations rule has same specificity as outer selector
PASS Nested declarations rule has top-level specificity behavior
PASS Nested declarations rule has top-level specificity behavior (max matching)
PASS Bare declartaion in nested grouping rule can match pseudo-element
PASS Nested group rules have top-level specificity behavior
FAIL Nested @scope rules behave like :where(:scope) assert_equals: expected "PASS" but got "FAIL"
FAIL Nested @scope rules behave like :where(:scope) (trailing) assert_equals: expected "PASS" but got "FAIL"
PASS Nested declarations rule responds to parent selector text change
